move ConcurrentReferenceHashMap back to Envers where it is actually used
